Refactor servers sagas
