+++ src/main/servers.js
@@ -1,6 +1,13 @@
+import fs from 'fs';
+import path from 'path';
 import url from 'url';
 
+import { app } from 'electron';
 import fetch from 'electron-main-fetch';
+import { select, call, put } from 'redux-saga/effects';
+
+import { PERSISTABLE_VALUES_MERGED } from '../actions';
+import { selectServers, selectCurrentServerUrl } from './selectors';
 
 export const ValidationResult = {
 	OK: Symbol('OK'),
@@ -70,3 +77,105 @@ export const normalizeServerUrl = (hostUrl) => {
 export const getServerInfo = async (/* serverUrl */) => {
 	throw Error('not implemented');
 };
+
+const loadAppServers = async () => {
+	try {
+		const filePath = path.join(
+			app.getAppPath(),
+			app.getAppPath().endsWith('app.asar') ? '..' : '.',
+			'servers.json',
+		);
+		const content = await fs.promises.readFile(filePath, 'utf8');
+		const json = JSON.parse(content);
+
+		return json && typeof json === 'object' ? json : {};
+	} catch (error) {
+		return {};
+	}
+};
+
+const loadUserServers = async () => {
+	try {
+		const filePath = path.join(app.getPath('userData'), 'servers.json');
+		const content = await fs.promises.readFile(filePath, 'utf8');
+		const json = JSON.parse(content);
+		await fs.promises.unlink(filePath);
+
+		return json && typeof json === 'object' ? json : {};
+	} catch (error) {
+		return {};
+	}
+};
+
+export function *setupServers(localStorage) {
+	let servers = yield select(selectServers);
+	let currentServerUrl = yield select(selectCurrentServerUrl);
+
+	const serversMap = new Map(
+		servers
+			.filter(Boolean)
+			.filter(({ url, title }) => typeof url === 'string' && typeof title === 'string')
+			.map((server) => [server.url, server]),
+	);
+
+	if (localStorage['rocket.chat.hosts']) {
+		try {
+			const storedString = JSON.parse(localStorage['rocket.chat.hosts']);
+
+			if (/^https?:\/\//.test(storedString)) {
+				serversMap.set(storedString, { url: storedString, title: storedString });
+			} else {
+				const storedValue = JSON.parse(storedString);
+
+				if (Array.isArray(storedValue)) {
+					storedValue.map((url) => url.replace(/\/$/, '')).forEach((url) => {
+						serversMap.set(url, { url, title: url });
+					});
+				}
+			}
+		} catch (error) {
+			console.warn(error);
+		}
+	}
+
+	if (serversMap.size === 0) {
+		const appConfiguration = yield call(loadAppServers);
+
+		for (const [title, url] of Object.entries(appConfiguration)) {
+			serversMap.set(url, { url, title });
+		}
+
+		const userConfiguration = yield call(loadUserServers);
+
+		for (const [title, url] of Object.entries(userConfiguration)) {
+			serversMap.set(url, { url, title });
+		}
+	}
+
+	if (localStorage['rocket.chat.currentHost'] && localStorage['rocket.chat.currentHost'] !== 'null') {
+		currentServerUrl = localStorage['rocket.chat.currentHost'];
+	}
+
+	servers = Array.from(serversMap.values());
+	currentServerUrl = serversMap.get(currentServerUrl)?.url ?? null;
+
+	if (localStorage['rocket.chat.sortOrder']) {
+		try {
+			const sorting = JSON.parse(localStorage['rocket.chat.sortOrder']);
+			if (Array.isArray(sorting)) {
+				servers = [...serversMap.entries()]
+					.sort(([a], [b]) => sorting.indexOf(a) - sorting.indexOf(b));
+			}
+		} catch (error) {
+			console.warn(error);
+		}
+	}
+
+	yield put({
+		type: PERSISTABLE_VALUES_MERGED,
+		payload: {
+			servers,
+			currentServerUrl,
+		},
+	});
+}
